Move job handling into module
